+++ lib/src/model/cache_options.dart
@@ -1,5 +1,4 @@
 import 'package:dio/dio.dart';
-import 'package:flutter/foundation.dart';
 import 'package:uuid/uuid.dart';
 
 import '../model/cache_priority.dart';
@@ -8,15 +7,15 @@ import '../store/cache_store.dart';
 /// Key builder to customize your keys.
 typedef CacheKeyBuilder = String Function(RequestOptions request);
 
-/// Encrypt content method.
+/// Encrypt content/headers method.
 typedef Encrypt = Future<List<int>> Function(List<int> bytes);
 
-/// Decrypt content method.
+/// Decrypt content/headers method.
 typedef Decrypt = Future<List<int>> Function(List<int> bytes);
 
 /// Policy to handle request behaviour.
 enum CachePolicy {
-  /// Forces to return the cached value if available.
+  /// Returns the cached value if available.
   /// Requests otherwise.
   cacheFirst,
 
@@ -29,38 +28,44 @@ enum CachePolicy {
 
   /// Requests and skips cache save even if
   /// response has cache directives.
+  ///
+  /// Note: previously stored response stays untouched.
   cacheStoreNo,
 
   /// Forces to request, even if a valid
-  /// cache is available.
+  /// cache is available and caches if
+  /// response has cache directives.
   refresh,
 
-  /// Requests and caches if
-  /// response has cache directives.
+  /// Requests and caches if response has directives.
   requestFirst,
 }
 
 /// Options to apply to handle request and cache behaviour.
 class CacheOptions {
-  /// Handle behaviour to request backend.
+  /// Handles behaviour to request backend.
   final CachePolicy policy;
 
   /// Ability to return cache excepted on given status codes.
-  /// Giving an empty list will hit cache on any error.
+  /// Giving an empty list will hit cache on any status codes.
+  ///
+  /// Other errors, such as socket exceptions (connect, send TO, receive TO,
+  /// ...),
+  /// will trigger the cache.
   final List<int> hitCacheOnErrorExcept;
 
   /// Builds the unique key used for indexing a request in cache.
   /// Default to [CacheOptions.defaultCacheKeyBuilder]
   final CacheKeyBuilder keyBuilder;
 
-  /// Override any HTTP directive to delete entry past this duration.
+  /// Overrides any HTTP directive to delete entry past this duration.
   final Duration maxStale;
 
   /// The priority of a cached value.
   /// Ease the clean up if needed.
   final CachePriority priority;
 
-  /// The store used for caching data.
+  /// Optional store used for caching data.
   final CacheStore store;
 
   /// Optional method to decrypt cache content
@@ -83,11 +88,10 @@ class CacheOptions {
     this.priority = CachePriority.normal,
     this.decrypt,
     this.encrypt,
-    @required this.store,
+    this.store,
   })  : assert(policy != null),
         assert(keyBuilder != null),
         assert(priority != null),
-        assert(store != null),
         assert((decrypt == null && encrypt == null) ||
             (decrypt != null && encrypt != null));