+++ lib/src/model/cache_response.dart
@@ -1,49 +1,70 @@
-import 'dart:convert';
-import 'dart:io';
+import 'dart:convert' show jsonDecode, utf8;
+import 'dart:io' show HttpStatus;
 
 import 'package:dio/dio.dart';
+import 'package:dio_cache_interceptor/src/model/cache_control.dart';
 import 'package:meta/meta.dart';
 
 import '../content_serialization.dart';
 import 'cache_priority.dart';
 
+/// Response representation from cache store.
 class CacheResponse {
   /// Cache key available in [Response]
   static const String cacheKey = '@cache_key@';
 
+  /// Available in [Response] if coming from this object.
+  static const String fromCache = '@fromCache@';
+
+  /// Response Cache-control header
+  final CacheControl cacheControl;
+
+  /// Response body
+  List<int> content;
+
+  /// Response Date header
+  final DateTime date;
+
+  /// ETag header
+  final String eTag;
+
+  /// Expires header
+  final DateTime expires;
+
+  /// Response headers
+  List<int> headers;
+
   /// Key used by store
   final String key;
 
-  /// Initial request URL
-  final String url;
-
-  /// Last modified header
+  /// Last-modified header
   final String lastModified;
 
   /// Max stale expiry
   final DateTime maxStale;
 
-  /// ETag header
-  final String eTag;
-
-  /// Response body
-  List<int> content;
-
-  /// Response headers
-  List<int> headers;
-
   /// Cache priority
   final CachePriority priority;
 
+  /// Absolute date representing date/time when response has been received
+  final DateTime responseDate;
+
+  /// Initial request URL
+  final String url;
+
   CacheResponse({
+    @required this.cacheControl,
+    @required this.content,
+    @required this.date,
+    @required this.eTag,
+    @required this.expires,
+    @required this.headers,
     @required this.key,
-    @required this.url,
     @required this.lastModified,
     @required this.maxStale,
-    @required this.eTag,
-    @required this.content,
-    @required this.headers,
     @required this.priority,
+    @required this.responseDate,
+    @required this.url,
   });
 
   /// Returns date in seconds since epoch or null.
@@ -57,13 +78,13 @@ class CacheResponse {
 
   Response toResponse(RequestOptions options) {
     final decHeaders = jsonDecode(utf8.decode(headers)) as Map<String, dynamic>;
-    final h = Headers();
 
+    final h = Headers();
     decHeaders.forEach((key, value) => h.set(key, value));
 
     return Response(
       data: deserializeContent(options.responseType, content),
-      extra: options.extra..addAll({cacheKey: key}),
+      extra: {cacheKey: key, fromCache: true},
       headers: h,
       statusCode: HttpStatus.notModified,
       request: options,