+++ lib/src/store/db_cache_store.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:dio_cache_interceptor/src/model/cache_control.dart';
 import 'package:path/path.dart';
 import 'package:sqflite/sqflite.dart';
 
@@ -7,19 +8,26 @@ import '../model/cache_priority.dart';
 import '../model/cache_response.dart';
 import 'cache_store.dart';
 
-/// A store that save each request result in a dedicated database.
+/// A store saving responses in a dedicated database
+/// from an optional [directory].
 ///
 class DbCacheStore extends CacheStore {
   static const String _tableName = 'dio_cache';
+
+  static const String _columnDate = 'date';
   static const String _columnKey = 'key';
+  static const String _columnCacheControl = 'cache_control';
   static const String _columnContent = 'content';
   static const String _columnETag = 'etag';
+  static const String _columnExpires = 'expires';
   static const String _columnHeaders = 'headers';
   static const String _columnLastModified = 'last_modified';
   static const String _columnMaxStale = 'max_stale';
   static const String _columnPriority = 'priority';
+  static const String _columnResponseDate = 'response_date';
   static const String _columnUrl = 'url';
-  static const int _currentDbVersion = 1;
+
+  static const int _currentDbVersion = 2;
 
   /// Database name. Optional.
   /// Useful if you want more than one DB.
@@ -36,13 +44,13 @@ class DbCacheStore extends CacheStore {
   Database _db;
 
   DbCacheStore({this.databaseName = _tableName, this.databasePath}) {
-    clean(stalledOnly: true);
+    clean(staleOnly: true);
   }
 
   @override
   Future<void> clean({
     CachePriority priorityOrBelow = CachePriority.high,
-    bool stalledOnly = false,
+    bool staleOnly = false,
   }) async {
     final db = await _getDatabase();
 
@@ -50,7 +58,7 @@ class DbCacheStore extends CacheStore {
       final where = StringBuffer(
         '$_columnPriority <= ${priorityOrBelow.index}',
       );
-      _whereMaxStale(stalledOnly, where);
+      _whereMaxStale(staleOnly, where);
 
       await db.delete(_tableName, where: where.toString());
     }
@@ -59,12 +67,12 @@ class DbCacheStore extends CacheStore {
   }
 
   @override
-  Future<void> delete(String key, {bool stalledOnly = false}) async {
+  Future<void> delete(String key, {bool staleOnly = false}) async {
     final db = await _getDatabase();
 
     if (db != null) {
       final where = StringBuffer('$_columnKey = \"$key\"');
-      _whereMaxStale(stalledOnly, where);
+      _whereMaxStale(staleOnly, where);
 
       await db.delete(_tableName, where: where.toString());
     }
@@ -110,15 +118,24 @@ class DbCacheStore extends CacheStore {
       }
 
       return CacheResponse(
-        key: key,
+        cacheControl:
+            CacheControl.fromHeader(result[_columnCacheControl]?.split(', ')),
         content: result[_columnContent],
+        date: result[_columnDate] != null
+            ? DateTime.tryParse(result[_columnDate])
+            : null,
         eTag: result[_columnETag],
+        expires: result[_columnExpires] != null
+            ? DateTime.tryParse(result[_columnExpires])
+            : null,
         headers: result[_columnHeaders],
+        key: key,
         lastModified: result[_columnLastModified],
         maxStale: maxStale != null
             ? DateTime.fromMillisecondsSinceEpoch(maxStale * 1000, isUtc: true)
             : null,
         priority: CachePriority.values[result[_columnPriority]],
+        responseDate: DateTime.tryParse(result[_columnResponseDate]),
         url: result[_columnUrl],
       );
     }
@@ -134,13 +151,17 @@ class DbCacheStore extends CacheStore {
       await db.insert(
         _tableName,
         {
-          _columnKey: response.key,
+          _columnDate: response.date.toIso8601String(),
+          _columnCacheControl: response.cacheControl?.toHeader(),
           _columnContent: response.content,
           _columnETag: response.eTag,
+          _columnExpires: response.expires?.toIso8601String(),
           _columnHeaders: response.headers,
+          _columnKey: response.key,
           _columnLastModified: response.lastModified,
           _columnMaxStale: response.getMaxStaleSeconds(),
           _columnPriority: response.priority.index,
+          _columnResponseDate: response.responseDate.toIso8601String(),
           _columnUrl: response.url,
         },
         conflictAlgorithm: ConflictAlgorithm.replace,
@@ -148,8 +169,8 @@ class DbCacheStore extends CacheStore {
     }
   }
 
-  void _whereMaxStale(bool stalledOnly, StringBuffer where) {
-    if (stalledOnly) {
+  void _whereMaxStale(bool staleOnly, StringBuffer where) {
+    if (staleOnly) {
       final expiry = DateTime.now().toUtc().millisecondsSinceEpoch / 1000;
 
       where.write(' AND ');
@@ -166,26 +187,44 @@ class DbCacheStore extends CacheStore {
         join(path, '$databaseName.db'),
         version: _currentDbVersion,
         onCreate: (Database db, int version) async {
-          await db.execute(_getCreateTableSql());
+          await db.execute(_getCreateTableSqlV2());
+        },
+        onUpgrade: (db, oldVersion, newVersion) async {
+          final batch = db.batch();
+          if (oldVersion == 1) {
+            _updateTableV1toV2(batch);
+          }
+          await batch.commit();
         },
       );
     }
     return _db;
   }
 
-  String _getCreateTableSql() {
+  String _getCreateTableSqlV2() {
     return '''
-      CREATE TABLE IF NOT EXISTS $_tableName ( 
-        $_columnKey text, 
+      CREATE TABLE IF NOT EXISTS $_tableName (
+        $_columnDate TEXT,
+        $_columnKey TEXT,
+        $_columnCacheControl TEXT,
         $_columnContent BLOB,
-        $_columnETag text,
+        $_columnETag TEXT,
+        $_columnExpires TEXT,
         $_columnHeaders BLOB,
-        $_columnLastModified text,
-        $_columnMaxStale integer,
-        $_columnPriority integer,
-        $_columnUrl text,
+        $_columnLastModified TEXT,
+        $_columnMaxStale INTEGER,
+        $_columnPriority INTEGER,
+        $_columnResponseDate TEXT,
+        $_columnUrl TEXT,
         PRIMARY KEY ($_columnKey)
         ) 
       ''';
+  }
+
+  void _updateTableV1toV2(Batch batch) {
+    batch.execute('ALTER TABLE $_tableName ADD $_columnCacheControl TEXT');
+    batch.execute('ALTER TABLE $_tableName ADD $_columnDate TEXT');
+    batch.execute('ALTER TABLE $_tableName ADD $_columnExpires TEXT');
+    batch.execute('ALTER TABLE $_tableName ADD $_columnResponseDate TEXT');
   }
 }